+++ README.md
@@ -64,9 +64,17 @@ A fun and useful pet project to keep track of personal notes on various web tech
 ** v1.0 **
 
 - Introduce React
-- Data Storage
+
+* TODO List **
+
+- [ ] Get JS Linting implemented
+- [ ] plan data Storage
 
 ** v1.1 **
+
+- Data Storage
+
+** v1.2 **
 
 - Frontend animations
