+++ source/components/Footer.jsx
@@ -1,5 +1,10 @@
 import React from 'react';
 
+/*
+TODO:
+- Build footer
+*/
+
 class Footer extends React.Component {
   render() {
     return <footer>Footer</footer>;
