+++ source/components/entries/Entries.jsx
@@ -3,6 +3,11 @@ import React from 'react';
 // -- Components
 import Entry from './Entry';
 
+/*
+TODO:
+- Iterate Entries
+*/
+
 class Entries extends React.Component {
   render() {
     return <div className="entries">
