+++ source/components/header/Intro.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+/*
+TODO:
+- Move content out of here
+- Ohh, and write the content
+*/
+
+class Intro extends React.Component {
+  render() {
+    return <section className="intro">
+          <h2>Picking Technologies</h2>
+          <p>Always keep learning. <span>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi vitae gravida enim. Integer eget odio pellentesque, aliquam sapien quis, aliquam mauris. Etiam a volutpat magna.</span></p>
+        </section>;
+  }
+}
+
+export default Intro;
