+++ source/components/header/ToTop.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+/*
+TODO:
+- Functional back to top
+- Display at right time
+*/
+
+class ToTop extends React.Component {
+  render() {
+    return <div className="totop">
+          To Top
+        </div>;
+  }
+}
+
+export default ToTop;
